requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0

